Reworked function completely
The previous iteration was too complicated.
Created the easy function to test env variables
