+++ functions/go/env-annotation/go.mod
@@ -1,0 +1,7 @@
+module github.com/GoogleContainerTools/kpt-functions-catalog/functions/go/env-annotation
+
+go 1.14
+
+require (
+	sigs.k8s.io/kustomize/kyaml v0.9.1
+)
