+++ functions/go/env-annotation/local-resource/data.yaml
