--- functions/go/templater/go.mod
@@ -1,14 +1,0 @@
-module github.com/GoogleContainerTools/kpt-functions-catalog/functions/go/templater
-
-go 1.14
-
-require (
-	github.com/Masterminds/goutils v1.1.0 // indirect
-	github.com/Masterminds/semver v1.5.0 // indirect
-	github.com/Masterminds/sprig v2.22.0+incompatible
-	github.com/huandu/xstrings v1.3.2 // indirect
-	github.com/imdario/mergo v0.3.10 // indirect
-	github.com/mitchellh/copystructure v1.0.0 // indirect
-	k8s.io/apimachinery v0.19.2
-	sigs.k8s.io/kustomize/kyaml v0.9.1
-)
