--- functions/go/templater/local-resource/example-use.yaml
@@ -1,14 +1,0 @@
-apiVersion: v1alpha1
-kind: Templater
-metadata:
-  annotations:
-    config.kubernetes.io/function: |
-      container:
-        image: gcr.io/kpt-functions/templater:stable
-data:
-  annotationSet: |
-    kind: AnnotationSetter
-    key: check-datestamp
-    value: %s
-  entrypoint: |
-    {{- $_ := KPipe .Items (list (KYFilter (list (YFilter (printf .Data.annotationSet (now | date "2006-01-02")))))) -}}
